feat(writer): create a blog
Allows users with writer roles to create a blog. in addition, this fixes Registration issue, by
adding new required field to allow api to create users based on the client specifying the role they
would like to create the user for. e.g. a user can be created as an admin, writer or editor. The
learner role has been removed as it served no purpose.

fix # 28